+++ private/pkg/git/git.go
@@ -15,8 +15,12 @@
 package git
 
 import (
+	"bufio"
+	"bytes"
 	"context"
+	"errors"
 	"regexp"
+	"strings"
 
 	"github.com/bufbuild/buf/private/pkg/app"
 	"github.com/bufbuild/buf/private/pkg/command"
@@ -24,6 +28,19 @@ import (
 	"github.com/bufbuild/buf/private/pkg/storage/storageos"
 	"github.com/bufbuild/buf/private/pkg/tracing"
 	"go.uber.org/zap"
+)
+
+const (
+	gitCommand      = "git"
+	gitOriginRemote = "origin"
+	tagsPrefix      = "refs/tags/"
+	headsPrefix     = "refs/heads/"
+)
+
+var (
+	// ErrRemoteNotFound is returned from GetRemote when the specified remote name is not
+	// found in the current git checkout.
+	ErrRemoteNotFound = errors.New("git remote not found")
 )
 
 // Name is a name identifiable by git.
@@ -174,9 +191,122 @@ type Remote interface {
 func GetRemote(
 	ctx context.Context,
 	runner command.Runner,
-	env map[string]string,
+	envContainer app.EnvContainer,
 	dir string,
 	name string,
 ) (Remote, error) {
-	return getRemote(ctx, runner, env, dir, name)
+	return getRemote(ctx, runner, envContainer, dir, name)
+}
+
+// CheckForUncommittedGitChanges checks if there are any uncommitted and/or unchecked
+// changes from git based on the given directory.
+func CheckForUncommittedGitChanges(
+	ctx context.Context,
+	runner command.Runner,
+	dir string,
+) ([]string, error) {
+	stdout := bytes.NewBuffer(nil)
+	stderr := bytes.NewBuffer(nil)
+	var modifiedFiles []string
+	// Unstaged changes
+	if err := runner.Run(
+		ctx,
+		gitCommand,
+		command.RunWithArgs("diff", "--name-only"),
+		command.RunWithStdout(stdout),
+		command.RunWithStderr(stderr),
+		command.RunWithDir(dir),
+	); err != nil {
+		return nil, err
+	}
+	modifiedFiles = append(modifiedFiles, getAllTrimmedLinesFromBuffer(stdout)...)
+
+	stdout = bytes.NewBuffer(nil)
+	stderr = bytes.NewBuffer(nil)
+	// Staged changes
+	if err := runner.Run(
+		ctx,
+		gitCommand,
+		command.RunWithArgs("diff", "--name-only", "--cached"),
+		command.RunWithStdout(stdout),
+		command.RunWithStderr(stderr),
+		command.RunWithDir(dir),
+	); err != nil {
+		return nil, err
+	}
+
+	modifiedFiles = append(modifiedFiles, getAllTrimmedLinesFromBuffer(stdout)...)
+	return modifiedFiles, nil
+}
+
+// GetCurrentHEADGitCommit returns the current HEAD commit based on the given directory.
+func GetCurrentHEADGitCommit(
+	ctx context.Context,
+	runner command.Runner,
+	dir string,
+) (string, error) {
+	stdout := bytes.NewBuffer(nil)
+	stderr := bytes.NewBuffer(nil)
+	if err := runner.Run(
+		ctx,
+		gitCommand,
+		command.RunWithArgs("rev-parse", "HEAD"),
+		command.RunWithStdout(stdout),
+		command.RunWithStderr(stderr),
+		command.RunWithDir(dir),
+	); err != nil {
+		return "", err
+	}
+	return strings.TrimSpace(stdout.String()), nil
+}
+
+// GetRefsForGitCommitAndRemote returns all refs pointing to a given commit based on the
+// given remote for the given directory. Querying the remote for refs information requires
+// passing the environment for permissions.
+func GetRefsForGitCommitAndRemote(
+	ctx context.Context,
+	runner command.Runner,
+	envContainer app.EnvContainer,
+	dir string,
+	remote string,
+	gitCommitSha string,
+) ([]string, error) {
+	stdout := bytes.NewBuffer(nil)
+	stderr := bytes.NewBuffer(nil)
+	if err := runner.Run(
+		ctx,
+		gitCommand,
+		command.RunWithArgs("ls-remote", "--heads", "--tags", "--refs", remote),
+		command.RunWithStdout(stdout),
+		command.RunWithStderr(stderr),
+		command.RunWithDir(dir),
+		command.RunWithEnv(app.EnvironMap(envContainer)),
+	); err != nil {
+		return nil, err
+	}
+	scanner := bufio.NewScanner(stdout)
+	var refs []string
+	for scanner.Scan() {
+		line := strings.TrimSpace(scanner.Text())
+		if ref, found := strings.CutPrefix(line, gitCommitSha); found {
+			ref = strings.TrimSpace(ref)
+			if tag, isTag := strings.CutPrefix(ref, tagsPrefix); isTag {
+				refs = append(refs, tag)
+				continue
+			}
+			if branch, isBranchHead := strings.CutPrefix(ref, headsPrefix); isBranchHead {
+				refs = append(refs, branch)
+			}
+		}
+	}
+	return refs, nil
+}
+
+func getAllTrimmedLinesFromBuffer(buffer *bytes.Buffer) []string {
+	scanner := bufio.NewScanner(buffer)
+	var lines []string
+	for scanner.Scan() {
+		lines = append(lines, strings.TrimSpace(scanner.Text()))
+	}
+	return lines
 }
